feat: add MAPOTFPolicy base class
Implemented a rough skeleton of MAPO's custom `Policy` class.

Note that `MAPOTrainer` will use `SyncSamplesOptimizer`, the default
for `build_trainer`, which simply collects `config["train_batch_size"]`
samples and calls `learn_on_batch` in the policy. This was tested by
collecting 1000 samples (timesteps of agent-environment interaction)
in the `Pendulum-v0` Gym environment (automated tests should be
implemented next).

The actor and critic losses are mostly placeholders, but verify that
we can backpropagate through the critic to calculate the deterministic
policy gradient.

Both actor and critic are optimized by the same tf optimizer; they
shall be separated later.

The environment model is not created yet.

Signed-off-by: Ângelo Lovatto <angelolovatto@gmail.com>